Refactor: RedisChannel 리팩토링 및 RedisService 주입
