+++ src/main/java/com/example/taste/common/constant/RedisChannel.java
@@ -1,5 +1,6 @@
-package com.example.taste.domain.notification.redis;
+package com.example.taste.common.constant;
 
 public class RedisChannel {
 	public static final String NOTIFICATION_CHANNEL = "notification";
+	public static final String MATCH_CHANNEL = "match";
 }
